Feat/did (#12)
* feat: add dependencies

* feat: lint and mocha config

* feat: add secp256k1 lib

* Triggers Workflow

* add unit test

* chore: add eslint check

* feat: update eslint and add husky

* feat: voting power smt builder

* feat: voting ticket builder

* feat: refactor codebase to generic utility classes to support multiple curves and generic functions/testing

* feat: babyjub curve support

* re-format data structure did

* change seed pda

* fix lint

* remove discriminator

* fix unit test

* restructure source code

* fix test:lib cmd

* update test cmd ci

* update packages path

---------
