+++ .github/workflows/mocha-test.yml
@@ -14,14 +14,14 @@ jobs:
       - name: Install dependencies and run tests
         run: |
           yarn install --frozen-lockfile
-          yarn run test && yarn run testlib
+          yarn test
 
       - name: Crypto experiment tests
         run: |
-          cd programs/crypto-experiment
+          cd packages/crypto_experiment
           cargo test
           cd ..
           cd ..
 
       - name: Eslint check
-        run: yarn run lint
+        run: yarn lint
