+++ Cargo.toml
@@ -1,7 +1,5 @@
 [workspace]
-members = [
-    "programs/*"
-]
+members = ["programs/*", "packages/*"]
 
 [profile.release]
 overflow-checks = true
