+++ circuits/voting/vote_verifier.circom
