+++ library/constants/IndividualVotingPowerConstants.ts
@@ -1,0 +1,1 @@
+export const MAX_TOTAL_VOTING_POWER = 1000000000000;
