+++ library/constants/SMTConstants.ts
@@ -1,0 +1,1 @@
+export const SMT_LEVEL = 10;
