+++ packages/crypto_experiment/Cargo.toml
@@ -1,0 +1,53 @@
+[package]
+name = "crypto_experiment"
+version = "0.1.0"
+description = "Jubjub curve testing"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib", "lib"]
+name = "crypto_experiment"
+
+[features]
+no-entrypoint = []
+no-idl = []
+no-log-ix-name = []
+cpi = ["no-entrypoint"]
+default = []
+
+[dependencies]
+babyjubjub-rs = "0.0.10"
+ark-serialize = "0.4.2"
+borsh = { version = "0.9.3", features = ["const-generics"] }
+ark-bn254 = { version = "0.3.0", default-features = false, features = ["curve"] }
+ark-groth16 = { git = "https://github.com/arkworks-rs/groth16", rev = "765817f", default-features = false }
+ark-ff = "0.3.0"
+ark-ec = "0.3.0"
+serde-json-wasm = { version = "0.4.0" }
+anyhow = "1.0"
+thiserror = "1.0"
+
+[dependencies.serde]
+version = "1"
+features = ["derive"]
+
+[dependencies.ff]
+version = "0.11"
+features = ["derive"]
+package = "ff_ce"
+
+[dependencies.num]
+version = "0.4"
+
+[dependencies.num-bigint]
+version = "0.4"
+features = ["rand"]
+
+[dependencies.num-traits]
+version = "0.2.8"
+
+[dependencies.poseidon-rs]
+version = "0.0.8"
+
+[dependencies.rand]
+version = "0.8"
