+++ packages/crypto_experiment/src/lib.rs
@@ -1,0 +1,539 @@
+extern crate rand;
+extern crate ff;
+pub mod groth16verifier_electron_lab;
+
+#[cfg(test)]
+mod tests {
+  use babyjubjub_rs::{PointProjective, Point};
+  use num_bigint::BigInt;
+  use ff::{Field, PrimeField, to_hex};
+  pub type Fr = poseidon_rs::Fr; // alias
+  use num_traits::Num;
+  use super::groth16verifier_electron_lab;
+
+  #[test]
+  fn test_point_addition1() {
+    let p: Point = Point {
+      x: Fr::from_str(
+        "17777552123799933955779906779655732241715742912184938656739573121738514868268",
+      )
+      .unwrap(),
+      y: Fr::from_str(
+        "2626589144620713026669568689430873010625803728049924121243784502389097019475",
+      )
+      .unwrap(),
+    };
+    let q: Point = Point {
+      x: Fr::from_str(
+        "16540640123574156134436876038791482806971768689494387082833631921987005038935",
+      )
+      .unwrap(),
+      y: Fr::from_str(
+        "20819045374670962167435360035096875258406992893633759881276124905556507972311",
+      )
+      .unwrap(),
+    };
+    let res = p.projective().add(&q.projective()).affine();
+    assert_eq!(
+      res.x,
+      Fr::from_str("7916061937171219682591368294088513039687205273691143098332585753343424131937")
+        .unwrap()
+    );
+    assert_eq!(
+      res.y,
+      Fr::from_str("14035240266687799601661095864649209771790948434046947201833777492504781204499")
+        .unwrap()
+    );
+
+    let decimal_str_from_prime_field_repr =
+      BigInt::from_str_radix(to_hex(&res.x).as_str(), 16).unwrap().to_str_radix(10);
+    let original_decimal_str =
+      String::from("7916061937171219682591368294088513039687205273691143098332585753343424131937");
+    assert_eq!(decimal_str_from_prime_field_repr, original_decimal_str);
+  }
+
+  #[test]
+  fn test_point_addition2() {
+    let p: PointProjective = PointProjective {
+      x: Fr::from_str(
+        "17777552123799933955779906779655732241715742912184938656739573121738514868268",
+      )
+      .unwrap(),
+      y: Fr::from_str(
+        "2626589144620713026669568689430873010625803728049924121243784502389097019475",
+      )
+      .unwrap(),
+      z: Fr::one(),
+    };
+    let q: PointProjective = PointProjective {
+      x: Fr::from_str(
+        "17777552123799933955779906779655732241715742912184938656739573121738514868268",
+      )
+      .unwrap(),
+      y: Fr::from_str(
+        "2626589144620713026669568689430873010625803728049924121243784502389097019475",
+      )
+      .unwrap(),
+      z: Fr::one(),
+    };
+    let res = p.add(&q).affine();
+    assert_eq!(
+      res.x,
+      Fr::from_str("6890855772600357754907169075114257697580319025794532037257385534741338397365")
+        .unwrap()
+    );
+    assert_eq!(
+      res.y,
+      Fr::from_str("4338620300185947561074059802482547481416142213883829469920100239455078257889")
+        .unwrap()
+    );
+
+    let decimal_str_from_prime_field_repr =
+      BigInt::from_str_radix(to_hex(&res.y).as_str(), 16).unwrap().to_str_radix(10);
+    let original_decimal_str =
+      String::from("4338620300185947561074059802482547481416142213883829469920100239455078257889");
+    assert_eq!(decimal_str_from_prime_field_repr, original_decimal_str);
+  }
+
+  #[test]
+  fn test_point_addition3() {
+    let p: Point = Point { x: Fr::from_str("0").unwrap(), y: Fr::from_str("1").unwrap() };
+    let q: Point = Point { x: Fr::from_str("0").unwrap(), y: Fr::from_str("1").unwrap() };
+    let res = p.projective().add(&q.projective()).affine();
+    assert_eq!(res.x, Fr::from_str("0").unwrap());
+    assert_eq!(res.y, Fr::from_str("1").unwrap());
+
+    let decimal_str_from_prime_field_repr =
+      BigInt::from_str_radix(to_hex(&res.y).as_str(), 16).unwrap().to_str_radix(10);
+    let original_decimal_str = String::from("1");
+    assert_eq!(decimal_str_from_prime_field_repr, original_decimal_str);
+  }
+
+  #[test]
+  fn test_valid_proof_snarkjs() {
+    let vkey_str = r#"
+    {
+      "protocol": "groth16",
+      "curve": "bn128",
+      "nPublic": 22,
+      "vk_alpha_1": [
+       "20491192805390485299153009773594534940189261866228447918068658471970481763042",
+       "9383485363053290200918347156157836566562967994039712273449902621266178545958",
+       "1"
+      ],
+      "vk_beta_2": [
+       [
+        "6375614351688725206403948262868962793625744043794305715222011528459656738731",
+        "4252822878758300859123897981450591353533073413197771768651442665752259397132"
+       ],
+       [
+        "10505242626370262277552901082094356697409835680220590971873171140371331206856",
+        "21847035105528745403288232691147584728191162732299865338377159692350059136679"
+       ],
+       [
+        "1",
+        "0"
+       ]
+      ],
+      "vk_gamma_2": [
+       [
+        "10857046999023057135944570762232829481370756359578518086990519993285655852781",
+        "11559732032986387107991004021392285783925812861821192530917403151452391805634"
+       ],
+       [
+        "8495653923123431417604973247489272438418190587263600148770280649306958101930",
+        "4082367875863433681332203403145435568316851327593401208105741076214120093531"
+       ],
+       [
+        "1",
+        "0"
+       ]
+      ],
+      "vk_delta_2": [
+       [
+        "10857046999023057135944570762232829481370756359578518086990519993285655852781",
+        "11559732032986387107991004021392285783925812861821192530917403151452391805634"
+       ],
+       [
+        "8495653923123431417604973247489272438418190587263600148770280649306958101930",
+        "4082367875863433681332203403145435568316851327593401208105741076214120093531"
+       ],
+       [
+        "1",
+        "0"
+       ]
+      ],
+      "vk_alphabeta_12": [
+       [
+        [
+         "2029413683389138792403550203267699914886160938906632433982220835551125967885",
+         "21072700047562757817161031222997517981543347628379360635925549008442030252106"
+        ],
+        [
+         "5940354580057074848093997050200682056184807770593307860589430076672439820312",
+         "12156638873931618554171829126792193045421052652279363021382169897324752428276"
+        ],
+        [
+         "7898200236362823042373859371574133993780991612861777490112507062703164551277",
+         "7074218545237549455313236346927434013100842096812539264420499035217050630853"
+        ]
+       ],
+       [
+        [
+         "7077479683546002997211712695946002074877511277312570035766170199895071832130",
+         "10093483419865920389913245021038182291233451549023025229112148274109565435465"
+        ],
+        [
+         "4595479056700221319381530156280926371456704509942304414423590385166031118820",
+         "19831328484489333784475432780421641293929726139240675179672856274388269393268"
+        ],
+        [
+         "11934129596455521040620786944827826205713621633706285934057045369193958244500",
+         "8037395052364110730298837004334506829870972346962140206007064471173334027475"
+        ]
+       ]
+      ],
+      "IC": [
+       [
+        "5119463127330496043780062159508696055106919940947179366264920102367742541539",
+        "12757871626179084362705360414145796635249954610398070229650184783625361400732",
+        "1"
+       ],
+       [
+        "13334957174441042136583447247729857938760317754111497593302174949298619918398",
+        "15542259091221413277943640510039284091117093748985619911666620686846766183198",
+        "1"
+       ],
+       [
+        "14859263770586916149318555848414982821228525421655012755329784067302870422707",
+        "3518745438396226633389119676710926556419068141277069596466208282738550349876",
+        "1"
+       ],
+       [
+        "4367075093963490052926888770158536987602338181453385314244291558213249374385",
+        "6993459062906060094708836305430447084526266551786247467251280308619712967275",
+        "1"
+       ],
+       [
+        "10297122630978290293056117281814363156831175327087913145189475554227696762703",
+        "17344235237091538241948506789101093460578984889300258645005325751375503416374",
+        "1"
+       ],
+       [
+        "462089204147685231392731594550210675090772645540309960344916342255010492244",
+        "12239470361965789506003908741803663843248717764178180223895359014705803580362",
+        "1"
+       ],
+       [
+        "8842352514162139345174625884893415757761439886758324211406664929244605000799",
+        "19097793878593374582291079993220531329807512962105140586884142877878188286498",
+        "1"
+       ],
+       [
+        "13143085576250627531525024226342299883660766082294354280041864233617763681721",
+        "14710108012128380438860774351606403685403057323666727061926500633229261612609",
+        "1"
+       ],
+       [
+        "8345502958300366917451228908661025729260771094084734432000423110288295457234",
+        "15215388166833928288165839991205630052945517805454970823292029008993530499761",
+        "1"
+       ],
+       [
+        "20261643798361960981678069722999612710517791079433765266185365779172528501493",
+        "15667509864806342217573379820249926475315899336549582743968277258347406130960",
+        "1"
+       ],
+       [
+        "19474544973606577118010830638559048719887760605377834384701778876594713431874",
+        "806296900065117055064545507335881999692284015588293908667154402621129229179",
+        "1"
+       ],
+       [
+        "7884816384308473291716936754594920412210521342783298951458925414410264629109",
+        "20999664295561056590156691647229336942312756386426846474572660501531541665026",
+        "1"
+       ],
+       [
+        "4786674039967449751243138076638934215846365965412650576693975427334190148781",
+        "11652091540111219240030545592604632726155231761222859305534163128665915815937",
+        "1"
+       ],
+       [
+        "2020117069757294399257760413256278033488869095846248684485439484462834412750",
+        "19153078723214468644493083578644928712318857176810497150585498394382991300523",
+        "1"
+       ],
+       [
+        "7971688315396436301896902829731532096498263872966667923562643405018506069181",
+        "1297603523957572374421962756484312855606503411516824110694370341762503461798",
+        "1"
+       ],
+       [
+        "18887021764889759538724388534598665067153960795873515676698559831761992332877",
+        "11468300373519662671873616057289389174727760959803639407420695204472936214277",
+        "1"
+       ],
+       [
+        "4209765541197875504465295766621885732744983469344848928092466449668354687782",
+        "10590640235240151223471982444552400697724036933743467677762219736943125903873",
+        "1"
+       ],
+       [
+        "18981797195078223403215917340032752175706483937903268224851007774194616072362",
+        "7251654456987014126869042187595844443688480477394419473455154401853177588347",
+        "1"
+       ],
+       [
+        "1039341654431518961656008315758139196038942625944072318366038759757612081229",
+        "16406090320490186081488972835532571616691715331368237470716561169030321836364",
+        "1"
+       ],
+       [
+        "20666545112157921235901197558899597179543605819464341874321163967612830854734",
+        "4499276055561623209036264983185881244167257306455799596556617613533967104408",
+        "1"
+       ],
+       [
+        "4427768844510325737141576455234302127130414292693211528299273804204775631181",
+        "7591688135424725853766434510942741849147334625526567937064148127029072876549",
+        "1"
+       ],
+       [
+        "15537356637086023366296827601549330999220101522171987719396513423349968217098",
+        "16594114982351387403779078346710741401708569050687331999642079556244263159262",
+        "1"
+       ],
+       [
+        "15801023490091337299764382241298022527322322751498444660671075717506684398155",
+        "7767390515679391409857553474972858781126383859225966842262291227120076247170",
+        "1"
+       ]
+      ]
+     }
+     "#;
+    let proof_str = r#"
+    {
+      "pi_a": [
+        "17553046687791719110332124013864893226002914495278790590412388607451616514674",
+        "19717245808197926285366599979351010477271594267254064140009658178646173342125",
+        "1"
+      ],
+      "pi_b": [
+        [
+          "14095503833206641116972976745526609696960467147506727866083787241531218164394",
+          "885895056654294190909579572679754298382699709618896401647870060530430232265"
+        ],
+        [
+          "3950182862386973698038022196553039651910441805764802670687222505563366374524",
+          "1421620647885069063125923525982524210380712733621312943791186146707341381768"
+        ],
+        [
+          "1",
+          "0"
+        ]
+      ],
+      "pi_c": [
+        "20122087925142908848873787887725990688240887837249908022936737708086953037892",
+        "21354007598428149831937656774454829506016485262908310346421644443605004952049",
+        "1"
+      ],
+      "protocol": "groth16",
+      "curve": "bn128"
+    }
+      "#;
+    let pub_input_str = r#"
+    [
+      "4780535511571899583326111294083506559176772238321226886650621394617793510453",
+      "17396300737976436341534849034519235354539937473721941136069856092261910884937",
+      "18141518717948068117699668904657576836863652629644680321338467938788841695584",
+      "1948916511357568535167361255534630038320300620845290566635973566807099028816",
+      "10958985638690938653992297889987206943780340749304381988881415837329256421392",
+      "13347947963607114282566249361865875710551528869491364787356025019059782640101",
+      "15164008403255484169185226550240674859896194830887780338195830396760545943229",
+      "19609714337629593690421537754401615530155806518585304284486971211623991867745",
+      "8433880027608613025737499839707513426037475484013004228251039599388427177858",
+      "2811290804372956832559552006250101727132497556732801642896094559437377596543",
+      "5825838144455249427161195733472035707908119645541186827332071679614179248174",
+      "14716529154556966649813046398803034433685823420696666490179533783157037522552",
+      "12209575374060234963268711827676282501914753179250737951282106531541117065297",
+      "8667751324134619965795293617486996032911949081766506400819470748496143781774",
+      "14815279858461596262650153397819649876537958460671607804897911799061339497453",
+      "6207080566309896789742309433525572194262857448292432499646495111686314482377",
+      "7397436025792035856118736173695980119779999324721505042540012809959970873856",
+      "9743500729874388828658762263472893464657081354522302561812981037177879455311",
+      "14369644761413768982540682636633790199068152059243377171214090000448312386018",
+      "19025513178172559287458055463415354495756214050337992155264527774838624293824",
+      "17130425939085723927077214623031336130483755042907093219794991672826067590042",
+      "18184263572126876537589943246509422776148082735062334245702431738855246518402"
+    ]
+      "#;
+    let vkey = groth16verifier_electron_lab::parse_verification_key(vkey_str.to_string()).unwrap();
+    let prepared_vkey = groth16verifier_electron_lab::get_prepared_verifying_key(vkey);
+    let res = groth16verifier_electron_lab::verify_proof(
+      prepared_vkey,
+      proof_str.to_string(),
+      pub_input_str.to_string(),
+    );
+    assert!(res.unwrap());
+  }
+}
+
+// Proof generated for voter 0:  {
+//   "pi_a": [
+//     "17553046687791719110332124013864893226002914495278790590412388607451616514674",
+//     "19717245808197926285366599979351010477271594267254064140009658178646173342125",
+//     "1"
+//   ],
+//   "pi_b": [
+//     [
+//       "14095503833206641116972976745526609696960467147506727866083787241531218164394",
+//       "885895056654294190909579572679754298382699709618896401647870060530430232265"
+//     ],
+//     [
+//       "3950182862386973698038022196553039651910441805764802670687222505563366374524",
+//       "1421620647885069063125923525982524210380712733621312943791186146707341381768"
+//     ],
+//     [
+//       "1",
+//       "0"
+//     ]
+//   ],
+//   "pi_c": [
+//     "20122087925142908848873787887725990688240887837249908022936737708086953037892",
+//     "21354007598428149831937656774454829506016485262908310346421644443605004952049",
+//     "1"
+//   ],
+//   "protocol": "groth16",
+//   "curve": "bn128"
+// }
+// Public signals =   [
+//   "4780535511571899583326111294083506559176772238321226886650621394617793510453",
+//   "17396300737976436341534849034519235354539937473721941136069856092261910884937",
+//   "18141518717948068117699668904657576836863652629644680321338467938788841695584",
+//   "1948916511357568535167361255534630038320300620845290566635973566807099028816",
+//   "10958985638690938653992297889987206943780340749304381988881415837329256421392",
+//   "13347947963607114282566249361865875710551528869491364787356025019059782640101",
+//   "15164008403255484169185226550240674859896194830887780338195830396760545943229",
+//   "19609714337629593690421537754401615530155806518585304284486971211623991867745",
+//   "8433880027608613025737499839707513426037475484013004228251039599388427177858",
+//   "2811290804372956832559552006250101727132497556732801642896094559437377596543",
+//   "5825838144455249427161195733472035707908119645541186827332071679614179248174",
+//   "14716529154556966649813046398803034433685823420696666490179533783157037522552",
+//   "12209575374060234963268711827676282501914753179250737951282106531541117065297",
+//   "8667751324134619965795293617486996032911949081766506400819470748496143781774",
+//   "14815279858461596262650153397819649876537958460671607804897911799061339497453",
+//   "6207080566309896789742309433525572194262857448292432499646495111686314482377",
+//   "7397436025792035856118736173695980119779999324721505042540012809959970873856",
+//   "9743500729874388828658762263472893464657081354522302561812981037177879455311",
+//   "14369644761413768982540682636633790199068152059243377171214090000448312386018",
+//   "19025513178172559287458055463415354495756214050337992155264527774838624293824",
+//   "17130425939085723927077214623031336130483755042907093219794991672826067590042",
+//   "18184263572126876537589943246509422776148082735062334245702431738855246518402"
+// ]
+// Proof verification result:  true
+// Proof generated for voter 1:  {
+//   "pi_a": [
+//     "2887008891292041859719809807197861959463743311916473569084862862949760152288",
+//     "13565832647771299355688722819992643706844701576758887893191727510494822572710",
+//     "1"
+//   ],
+//   "pi_b": [
+//     [
+//       "13676922235429983493492632286672843949840095299524575722905043129261878315473",
+//       "2263469836299495155613693025227856952512469599341988520479739620374973683836"
+//     ],
+//     [
+//       "17373496169283419155091366005664189124383094308529712611161324950734655997640",
+//       "17153839384823950999283993982036147683886904128355309274557913439451998380410"
+//     ],
+//     [
+//       "1",
+//       "0"
+//     ]
+//   ],
+//   "pi_c": [
+//     "21152942003772921328895319116886404345001551005416242117251837432572329169085",
+//     "2284893786969492012138539378276842047144699028784825564496642353565718782433",
+//     "1"
+//   ],
+//   "protocol": "groth16",
+//   "curve": "bn128"
+// }
+// Public signals =   [
+//   "4780535511571899583326111294083506559176772238321226886650621394617793510453",
+//   "17396300737976436341534849034519235354539937473721941136069856092261910884937",
+//   "21677416414048002885014831617359018636762654537326105006796099190495638486779",
+//   "5698639305858714401986760868147899698655855046981359719434670336687417169054",
+//   "15757512298620128830007760698436855746078219409216006650815139484572348922503",
+//   "15395299098600682892238742308361023373874408300004693409774276033411433834650",
+//   "17792081549194100206224270627551984504594179778738807776425268862951089776362",
+//   "15355779618148005876044262644433673958943473483483929962463262452554486572561",
+//   "20982576201697614198020056431557971530716676702637523511351878626519296478723",
+//   "11897343908593337041309908666565054523123319379596153239986233946362652186154",
+//   "17643159566763501850891187562924927037748680580199471747038557483962648883508",
+//   "14168221663630227733738292076695383343993716864538827919142036152430552809442",
+//   "9170361922561136322591268077706439999663396281439960256135202761110273430892",
+//   "8285833560635393997947718756648734371469893044903646887478161099824639551261",
+//   "17449163393244189974846005775661569704884578861416455846572058778324088008091",
+//   "15399131125037571981436791496053738770407581824042125142269079900789763398063",
+//   "7718484908677169102767112963767559576801781012435886396657382284850383389882",
+//   "10236472788417205880404691858672802391933469562362640818319788222743215569213",
+//   "19258969294284990409034343209891752268759011395403226962535271346119672422514",
+//   "13412059844248031455324784789827110530810881806100463168045639027526328045990",
+//   "3709940303546077096286398930111638070657838786747619948480888528950630326991",
+//   "9466136737114432817225731137783092815453236395189419532008146435003922160614"
+// ]
+// Proof verification result:  true
+// Proof generated for voter 2:  {
+//   "pi_a": [
+//     "4981581316017416010968185124528551272927338819822849894940868676510331141615",
+//     "13172624465748521411191484510557674318189103400533154036073634978354523952789",
+//     "1"
+//   ],
+//   "pi_b": [
+//     [
+//       "13934974041411199452065669187466346621154488858850399021112880389430890033110",
+//       "15020885609179340036925032580425031532497732355961843159616452955658529331887"
+//     ],
+//     [
+//       "1704620069583348989391434205245310770445268347580160504631849328389565652310",
+//       "16199187963422846132585717884967175847366007334456107155391335529195386801335"
+//     ],
+//     [
+//       "1",
+//       "0"
+//     ]
+//   ],
+//   "pi_c": [
+//     "15140365468861432762074674483898629679224423670253986614282018883313512149983",
+//     "9878577008028596256553338620693815365247463107700306958678145676746248347568",
+//     "1"
+//   ],
+//   "protocol": "groth16",
+//   "curve": "bn128"
+// }
+// Public signals =   [
+//   "4780535511571899583326111294083506559176772238321226886650621394617793510453",
+//   "17396300737976436341534849034519235354539937473721941136069856092261910884937",
+//   "5999897246605649565507953420349949926067012258028404565337295241160895648529",
+//   "5188602884268676601547865209597981841765189576750672378427727267441549584528",
+//   "10504936643286583426567527183070663320328549091703463088053239037677412379781",
+//   "4608215392869310197022725944575186210897270799003030822684646841014456368679",
+//   "17872153610924921838975309015515583897055939816789481768192223907273121055154",
+//   "2133438170774239962078052577402124277269892860125430514347746551108544866433",
+//   "20151496780975238322906441840163696516593279469634276655292555176992702493772",
+//   "14652088676891529403930164322615880127304095992264096121592895938833313936500",
+//   "17169775100955531442102418106538287585589151483339926458728647828531185057226",
+//   "19323946407730810341146578132352244716110768124673214309541007601249683327360",
+//   "16939517714325738216371245828476147532255499839398211331787581727311551415328",
+//   "17799836023385595082072628683154593081601099853519226053526760054823634187171",
+//   "8671831495355022689104155000537824972998578935596246065173995991125324219710",
+//   "4945944967842468585930035660736534964457442479453001115988698817027182602809",
+//   "20540430034374955680364539564747901373528802695861128392976319388252300769002",
+//   "16718386766526628254754697655448392075425579606887946018715863291728506816297",
+//   "7584779268429220181575334236444408786001699141703535744580152617695300884989",
+//   "4452722596896297805430609850605349263236440633962528798434114290241221379478",
+//   "13777055453259574421469183514649215552918762934391675724722634205213089994423",
+//   "10249677223940046747629562286214807753147872652976340921290790545336232025213"
+// ]
